requirements.txt
absl-py==0.2.2
astor==0.7.1
certifi==2018.4.16
chardet==3.0.4
coremltools==0.8
cycler==0.10.0
gast==0.2.0
grpcio==1.13.0
idna==2.7
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.5
pandas==0.23.3
Pillow==5.2.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
requests==2.19.1
s3cmd==2.0.1
six==1.10.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
tfcoreml==0.2.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
